+++ api/CHANGELOG.md
@@ -46,6 +46,7 @@
 - Made saved patched app entries in the `Apps` tab not overwrite each other unless the app has the same package name and was patched with the same patch bundle
 - Added the ability to set a image of choice as the app background
 - Added `Always create new saved app entry` toggle in Settings > Advanced that toggles saved patch app entries from being overwritten
+- Added `Hide main tab labels` toggle in Settings > General that toggles the labels under the tab icons on the main screen
 
 
 # Bug fixes